requirements.txt
undetected-chromedriver>=3.5.4
selenium>=4.15.0
requests>=2.31.0
python-dotenv>=1.0.0
setuptools>=65.0.0
